requirements.txt
# Sakshi.AI - Intelligent Video Analytics Platform
# Clean, minimal requirements.txt with only essential dependencies

# ============================================================================
# Core Web Framework
# ============================================================================
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0

# ============================================================================
# Database
# ============================================================================
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7  # PostgreSQL driver (required for PostgreSQL)

# ============================================================================
# Real-time Communication (Socket.IO)
# ============================================================================
python-socketio==5.9.0
python-engineio==4.12.3
eventlet==0.33.3
simple-websocket==1.1.0
wsproto==1.2.0
h11==0.16.0
greenlet==3.2.4

# ============================================================================
# Computer Vision & AI (Core Dependencies)
# ============================================================================
opencv-python==4.8.1.78
numpy>=1.26.0,<2.0.0
# PyTorch - Install CUDA version separately for GPU support:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# For CPU-only: pip install torch torchvision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
Pillow>=10.0.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.0
PyYAML>=6.0.0
tqdm>=4.65.0
psutil>=5.9.0
filelock>=3.12.0

# ============================================================================
# Date/Time
# ============================================================================
python-dateutil>=2.8.0
pytz>=2023.3

# ============================================================================
# Network & HTTP (for API calls if needed)
# ============================================================================
requests>=2.31.0
urllib3>=2.0.0

# ============================================================================
# Core Dependencies (auto-installed, but listed for clarity)
# ============================================================================
# These are typically auto-installed by the packages above:
# - typing-extensions
# - packaging
# - six
# - certifi
# - charset-normalizer
# - idna

# ============================================================================
# Object Tracking
# ============================================================================
deep-sort-realtime==1.3.2  # Required for ServiceDisciplineMonitor

# ============================================================================
# Optional Dependencies (uncomment if needed)
# ============================================================================
# pandas>=2.0.0  # Data analysis (if generating reports)
# matplotlib>=3.7.0  # Visualization (if generating charts)
# openpyxl>=3.1.0  # Excel export (if generating Excel reports)

# ============================================================================
# GPU Support (Optional)
# ============================================================================
# For CUDA-enabled PyTorch, install separately:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# ============================================================================
# Notes:
# ============================================================================
# - torch and torchvision versions should match (check PyTorch compatibility)
# - ultralytics handles YOLO model loading and inference
# - opencv-python provides video processing capabilities
# - eventlet is used for async Socket.IO communication
# - All other dependencies are minimal and essential for core functionality
